refactor: monitorDistanceAndManagePlayStatus함수 viewModel로 이동
