+++ android/app/src/main/java/com/happy/friendogly/remote/api/ApiClient.kt
@@ -2,16 +2,14 @@ package com.happy.friendogly.remote.api
 
 class ApiClient {
     object Auth {
-        //        private const val BASE_URL = "/api/auth"
-        private const val BASE_URL = "/auth"
+        private const val BASE_URL = "/api/auth"
         const val POST_KAKAO_LOGIN = "$BASE_URL/kakao/login"
         const val POST_REFRESH = "$BASE_URL/kakao/refresh"
         const val POST_LOGOUT = "$BASE_URL/kakao/logout"
     }
 
     object PlayGround {
-        //        private const val BASE_URL = "/api/playgrounds"
-        private const val BASE_URL = "/playgrounds"
+        private const val BASE_URL = "/api/playgrounds"
         const val POST_PLAYGROUND = BASE_URL
         const val PATCH_PLAYGROUND_ARRIVAL = "$BASE_URL/arrival"
         const val GET_PLAYGROUNDS = "$BASE_URL/locations"
@@ -23,8 +21,7 @@ class ApiClient {
     }
 
     object Member {
-        //        private const val BASE_URL = "/api/members"
-        private const val BASE_URL = "/members"
+        private const val BASE_URL = "/api/members"
         const val POST_MEMBER = BASE_URL
         const val GET_MEMBER_MINE = "$BASE_URL/mine"
         const val GET_MEMBER = "$BASE_URL/{id}"
@@ -33,8 +30,7 @@ class ApiClient {
     }
 
     object Pet {
-        //        private const val BASE_URL = "/api/pets"
-        private const val BASE_URL = "/pets"
+        private const val BASE_URL = "/api/pets"
         const val GET_PETS_MINE = "$BASE_URL/mine"
         const val POST_PET = BASE_URL
         const val GET_PETS = BASE_URL
@@ -43,8 +39,7 @@ class ApiClient {
     }
 
     object Club {
-        //        private const val BASE_URL = "/api/clubs"
-        private const val BASE_URL = "/clubs"
+        private const val BASE_URL = "/api/clubs"
         private const val MEMBER_URL = "/members"
         const val POST_CLUB = BASE_URL
         const val GET_CLUB_SEARCHING = "$BASE_URL/searching"
@@ -55,15 +50,13 @@ class ApiClient {
     }
 
     object MyClub {
-        //        private const val BASE_URL = "/api/clubs"
-        private const val BASE_URL = "/clubs"
+        private const val BASE_URL = "/api/clubs"
         const val OWNING = "$BASE_URL/owning"
         const val PARTICIPATING = "$BASE_URL/participating"
     }
 
     object ChatRoom {
-        //        private const val BASE_URL = "/api/chat-rooms"
-        private const val BASE_URL = "/chat-rooms"
+        private const val BASE_URL = "/api/chat-rooms"
         const val CHAT_LIST = "$BASE_URL/mine"
         const val MEMBERS = "$BASE_URL/{chatRoomId}"
         const val CLUB = "$BASE_URL/{chatRoomId}/club"
@@ -71,8 +64,7 @@ class ApiClient {
     }
 
     object ChatMessage {
-        //        private const val BASE_URL = "/api/chat-messages"
-        private const val BASE_URL = "/chat-messages"
+        private const val BASE_URL = "/api/chat-messages"
         const val ALL = "$BASE_URL/{chatRoomId}"
         const val TIMES = "$BASE_URL/{chatRoomId}/times"
     }
